+++ lionagi/service/providers/openai_/chat_completions.py
